+++ src/commands/manuals/test.rs
@@ -80,7 +80,7 @@ OPTIONS
     --fileset <key=glob>...
         A glob-style pattern identified by name to include in the blueprint
 
-    --no-clean
+    --keep, -k
         Do not clean the target folder before execution
 
     --force